# api/docs/swagger.yaml
definitions:
...
  memory.ShareCreateBody:
    properties:
      memory_id:
        type: string
      shared_with:
        items:
          type: string
        type: array
    type: object
  memory.ShareDeleteBody:
    properties:
      memory_id:
        type: string
      unshared_with:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
  memory.ShareRes:
    properties:
      memony_title:
        type: string
      shared_with:
        items:
          type: string
        type: array
    type: object
  timeline.GetAllRes:
    properties:
      millistones:
        items:
          $ref: '#/definitions/timeline.Millistones'
        type: array
      total_count:
        type: integer
    type: object
  timeline.Millistone:
    properties:
      category:
        type: string
      create_at:
        type: string
      date:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      id:
        type: string
      title:
        type: string
      update_at:
        type: string
      user_id:
        type: string
    type: object
  timeline.Millistones:
    properties:
      category:
        type: string
      create_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  timeline.MillistonesCreateBody:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Time Capsule resources
  title: Time Capsule API
  version: "1.0"
paths:
...
  /millistones:
    get:
      consumes:
      - application/json
      description: Retrieve a list of millistones based on optional filters like date,
        category, user ID, limit, and offset
      parameters:
      - description: Date filter (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: Category filter
        in: query
        name: category
        type: string
      - description: User ID filter
        in: query
        name: user_id
        type: string
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset number for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timeline.GetAllRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all millistones based on filters
      tags:
      - millistones
    post:
      consumes:
      - application/json
      description: Create a new millistone with the specified details
      parameters:
      - description: Millistoneobjecttobecreated
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/timeline.MillistonesCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Millistone created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new millistone
      tags:
      - millistones
  /millistones/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a millistone by its ID
      parameters:
      - description: Millistone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Millistone deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a millistone by ID
      tags:
      - millistones
    get:
      consumes:
      - application/json
      description: Retrieve details of a millistone by its ID
      parameters:
      - description: Millistone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timeline.Millistone'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a millistone by ID
      tags:
      - millistones
    put:
      consumes:
      - application/json
      description: Update an existing millistone with the specified details
      parameters:
      - description: Millistone ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated millistone details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/timeline.MillistonesCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Millistone updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing millistone
      tags:
      - millistones
  /millistones/date:
    get:
      consumes:
      - application/json
      description: Retrieve millistones within a specified date range for a given
        user ID
      parameters:
      - description: Start date of range (YYYY-MM-DD)
        in: query
        name: from_date
        required: true
        type: string
      - description: End date of range (YYYY-MM-DD)
        in: query
        name: to_date
        required: true
        type: string
      - description: User ID filter
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timeline.GetAllRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get millistones within a date range
      tags:
      - millistones
  /shares:
    get:
      consumes:
      - application/json
      description: Get all shares for a memory
      parameters:
      - description: Memory ID
        in: query
        name: memory_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.ShareRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get shares
      tags:
      - shares
    post:
      consumes:
      - application/json
      description: Share a memory with other users
      parameters:
      - description: Share Create Body
        in: body
        name: share
        required: true
        schema:
          $ref: '#/definitions/memory.ShareCreateBody'
      produces:
      - application/json
      responses:
        "201":
          description: Share created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a share
      tags:
      - shares
    put:
      consumes:
      - application/json
      description: Remove a share for a memory
      parameters:
      - description: Share ID
        in: query
        name: share_id
        required: true
        type: string
      - description: Share Delete Body
        in: body
        name: share
        required: true
        schema:
          $ref: '#/definitions/memory.ShareDeleteBody'
      produces:
      - application/json
      responses:
        "200":
          description: Share updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Unshare a memory
      tags:
      - shares
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"